quiz_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
python-multipart
kafka-python
boto3